backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
faiss-cpu==1.9.0.post1
google-generativeai==0.8.4
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.20
